Improved colSpan calculations and altered DOM classes
data-es6 (and the rest) now record the platformtype of each platform - "desktop" (default if absent), "mobile", "compiler" and "engine". This is used as a class for each platform's <th> element, and is used by master.js to dynamically assign all of the platformtype <th> elements' colSpans.

platformtype <th> elements now have an ID corresponding to their identity, and platform <th> elements now have the "platform" class.