Annex B features are now weighted to 1/5 for percent calculation.
By treating the entire Annex B suite of features as a single "feature", the percentages for very early platforms such as IE 10 are more accurate. Weighting other feature sets will not be done, though, because it'd be too complicated and subjective deciding whether e.g. destructuring is "worth" the same % as typed arrays.

Also added a legend to the ES6 table regarding the platform bars' colours.