Made iOS7/8 and Phantom results strictly equal Safari results.
As iOS7/8 and Phantom have identical JS engines to Safari 7, 8 and 6 respectively, I've now allowed an "equals" property to be used in data-es6.js to allow their results to be defined in terms of the 'parent' Safari versions. Thus, specifying results for these platforms individually is no longer necessary.

Also, closed #530.