iojs results are now expressed in terms of Chrome 44.
As it turns out, iojs and Chrome 44 have the same V8 version, and testing has shown their results to be identical. So, they are now expressed as such in data-es6.js.

This also fixes a few small issues with the Chrome 44 data.

Also fixed a CSS bug that caused the horizontal scrollbar to be missing.